requirements.txt
requests==2.25.1
python-dotenv==0.17.0
pytz==2021.3